+++ CHANGELOG.md
@@ -122,6 +122,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
     - DerefMut for Ber
     - 'TryFrom<&mut [u8]>' for '&mut BerRef'
     - 'TryFrom<&mut [u8]>' for '&mut DerRef'
+    - 'AsRef<IdRef>' for Id
+    - 'AsMut<IdRef>' for Id
     - 'PartialEq<T>' for 'Id' where 'T' is bounded on 'Borrow<IdRef>'
     - 'PartialOrd<T>' for 'Id' where 'T' is bounded on 'Borrow<IdRef>'
     - 'PartialEq<T>' for 'IdRef' where 'T' is bounded on 'Borrow<IdRef>'
@@ -161,6 +163,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
     - IdRef.as\_mut\_bytes()
 ### Removed
 - Delete the implementations for the following traits
+    - Boffow<[u8]> for Id
     - Boffow<[u8]> for IdRef
     - Deref for IdRef
     - DerefMut for IdRef